Update test_loadCombination.py
Since we are running through all possible Standards in RFEM, we need to change the LoadCase assertions to be for an ActionCategoryType which is common among all Standards. This explains the changes to test_loadCombination.py file.

Initial failing test was found in test_loadCasesAndCombinations.py. However, within this file, the test_loadCombination() function is called within a try and except block.